move chart disposable
