+++ src/components/simulator/input/d3Chart/disposable/D3ChartDisposable.tsx
